pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llx"
version = "0.1.8"
description = "Intelligent LLM model router driven by real code metrics — successor to preLLM"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{name = "Tom Sapletta", email = "tom@sapletta.com"}]
keywords = ["llm", "code-analysis", "model-routing", "litellm", "code2llm", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "tomli>=2.0; python_version<'3.11'",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "requests>=2.31",
    "docker>=6.0",
    "psutil>=5.9",
]

[project.scripts]
llx = "llx.cli.app:main"
llx-tools = "llx.tools.cli:main"
llx-orchestrator = "llx.orchestration.orchestrator_cli:main"
llx-mcp = "llx.mcp.server:main_sync"

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
litellm = ["litellm>=1.40"]
code2llm = ["code2llm>=0.9"]
redup = ["redup>=0.4"]
vallm = ["vallm>=0.1"]
ollama = ["ollama>=0.3"]
prellm = [
    "litellm>=1.40",
    "jinja2>=3.1",
    "nfo>=0.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]
prellm-full = [
    "litellm>=1.40",
    "jinja2>=3.1",
    "nfo>=0.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "chromadb>=0.4",
    "tree-sitter>=0.21",
    "instructor>=1.0",
]
all = [
    "litellm>=1.40",
    "code2llm>=0.9",
    "redup>=0.4",
    "vallm>=0.1",
    "ollama>=0.3",
    "mcp>=1.0",
    "jinja2>=3.1",
    "nfo>=0.1",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.urls]
Homepage = "https://github.com/wronai/llx"
Repository = "https://github.com/wronai/llx"

[tool.setuptools.packages.find]
where = ["."]
include = ["llx*"]

[tool.setuptools.package-data]
"llx" = ["configs/**/*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
